requirements.txt
amqp==2.5.2
apns2==0.7.1
argon2-cffi==19.2.0
asgiref==3.2.7
astroid==2.4.1
Babel==2.8.0
bcrypt==3.1.7
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-celery==3.1.17
django-cors-headers==3.2.1
django-filter==2.2.0
django-phonenumber-field==4.0.0
django-push-notifications==2.0.0
Django-Pushy==1.0.2
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
drf-writable-nested==0.6.0
drf-yasg==1.17.1
future==0.18.2
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.9
inflection==0.4.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
joblib==0.14.1
kombu==4.6.8
lazy-object-proxy==1.4.3
livereload==2.6.1
lunr==0.5.6
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
mkdocs==1.1
mysqlclient==1.4.6
nltk==3.5
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
phonenumbers==8.12.2
Pillow==7.1.1
pushjack==1.3.0
py-vapid==1.7.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
pyparsing==2.4.7
python3-openid==3.1.0
pytz==2019.3
pywebpush==1.11.0
PyYAML==5.3.1
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
tornado==6.0.4
tqdm==4.45.0
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
wincertstore==0.2
wrapt==1.12.1
